requirements.txt
torch>=1.7.0
torchvision>=0.8.1
Pillow>=8.0.0
transformers>=4.0.0
numpy==1.23.5